feat: Switching to A1h SAT command by default
After looking at lots of USB devices, HBAs, and other OSs, the A1h operation code for SAT ATA-Passthrough seems to be supported by far more devices than 85h. Because of this and the other weirdness associated with USB devices, it seems better to use A1h when it can be used as it is widely supported.
Since some of these changes were far reaching, I also went ahead and added support for issuing 32B SAT CDBs as well.
I removed one hack since it is no longer necessary with A1h being the default operation code to use at this point. It was also a little redundant due to having another hack that can be set when A1h is not supported.