doc: Adding comment with potential hack for 48bit via A1h SAT CDB
There are a few devices that only support the A1h opcode, but are otherwise drives that support GPL or other 48bit commands. Because it may be possible to get an ext command through with this in some situations, I added commented out code that MAY help some of these devices if others are encountered. This is currently commented out because it did not work on the one device that I have in my possession.