feat: Adding ASMedia USB to NVMe support
Added support for ASMedia USB to NVMe passthrough based on documentation.
There is a basic passthrough which is limited to identify and get log page and a full passthrough which doesn't seem to be supported on all devices.
Due to adding this support, additional NVMe specific hacks are added to help handle the limitations.