doc: Adding a comment about ASMedia NVMe packet passthrough for current bug
There is currently a bug for ASMedia NVMe packet passthrough.
I will be receiving a development device which has this capability active in the chip which will enable me to debug this better and figure out what is wrong.
Currently, the data phase is being rejected with "invalid field in CDB", so I need a device to play with and adjust the CDB in the debugger until I get it right.
I have added the new device in the passthrough hacks based on initial remote testing and other documentation I have received for it.