bug: Fixing ASMedia NVMe passthrough packet issues
Now that I have a device in hand, I was able to debug and completely figure out all of the issues with the ASMedia Packet passthrough.
There were a few bugs in validating parameters when creating the CDBs as well as needing to handle transfers that were not 512B aligned.
Once these were resolved and tested, I found a problem in the hack lookup table for the specific device that I have and removed a couple hacks that were unnecessarily in place.