feat: Adding functions for OS level resets
Adding functions to do OS level SAS/SATA device, bus, and controller resets.
This is not implmeneted on all OSs since not all have a method to do this.
Some OSs support more resets than others.
In some cases, low-level drivers may need to support the calls in addition to the OS itself for it to work.