bug: Adding USB to SAS adapter to lookup table
Adding a USB to SAS adapter to the internal USB lookup table. This thing kinda works, but lots of things don't.
In order to help it out, some hacks were set to force no ATA passthrough testing, which seems to have helped a little.
Many commands still don't go through, but as far as I can tell, this is an adapter limitation.