hot: Additional condition before disabling SAT check
Adding in one additional check to make sure that the low-level USB device ID wasn't already detected and setup for SAT support. This should help in case a USB adapter sets a legacy bit that is not expected or something else that would normally disable a SAT check shows up.