bug: Adding hacks for VS2013
Adding some hacks to workaround how scanf macros were defined for VS2013.
Because of how they were defined, the TEXT macro will not work on them, so needed to work around this to make VS2013 capable of compiling the code still.
VS2015 and newer do not have this issue, so they continue using the macros.