bug: Removing packing from identify data to help reduce warnings
Removing packing instructions from ATA and NVMe identify data to help clean up warnings and prevent possible issues on other processor types.
While the packing was necessary, cleaning up the ATA identify structure and making sure it is aligned in a device structure should have it work the same.
Checking NVMe devices shows it works without packing as well, likely due to the identify data being layed out with this kind of thing in mind as NVMe is being defined in a spec committee.