feat: Updating suppored NVMe commands based on MSFT documentation
Based on newly found documentation, updating the supported NVMe commands in Windows.
Some commands are only available in Windows PE.
It is unclear which version/update of Win10 allows access to some of these commands at this time so it is not perfect. More testing and debugging will be required to narrow it down further, but this is better than the guessing and trial and error that was used before.