feat: Adding Intel RST unique IOCTLs. Focussed on FWDL and NVMe passthrough
Adding the first code to send some Intel RST driver unique IOCTLs. This has not yet been tested and I need to debug it on actual hardware, so it isn't being used yet.
These IOCTLs work in conjunction with CSMI and I need to make changes to CSMI discovery to detect when these are actually available to use.
Waiting on access to hardware setup for these IOCTLs before I push any changes to the CSMI code since I need to debug and see how these devices will show up and report before I use them. Since Intel may not be the only driver implementing CSMI, we need to be careful to not break other implementations.