bug: Fixing some issues in getting list of devices in sg_helper
Fixing a couple of bugs in sg_helper related to getting the device list.
If not run as root, but the user has access to some device types (such as cdrom), then we were segmentation faulting.
Also fixing not properly returning permission denied errors so that they could be caught in the functions that call get_Device_List