feat: Massive changes to CSMI to support Intel RST addons
Making massive changes to rewrite CSMI support to support Windows and Linux properly, as well as changes to support the ability to use the Intel RST additions to issue additional IOCTLs for NVMe passthrough and firmware download.

While opensea-transport is building, including files from it results in errors for opensea-operations. Still working on it, but need to check in what I have so far so I can begin testing.