feat: Adding raid scan helper to assist with scanning RAID interfaces
Adding a linked list called raid_scan_helper to assist with scanning RAID interfaces in a given OS.
The list holds a string "handle" and hints that the OS can provide to the RAID lib scanning functions to hint at what it thinks may be a certain type of raid, or even "unknown".
This is useful because it will make it easy to add various methods to scan different kinds of RAIDs. How an OS guesses which RAID it is is OS dependent, but it can be looking at drive information, driver being used, etc to make the scan better as more interfaces are added.
Currently works in CSMI with Intel RST additions. More RAIDs to be added later.