feat: Adding lookup for security access and CSMI Info output
Adding in code to support checking the Windows registry to determine CSMI IOCTL access since this can be implemented by a driver.
Also adding support for printing out some of the CSMI information which can be helpful with debugging. This information concentrates on the things being stored in the CSMI structure since that is used to issue commands.