bug: Fixing some known big endian issues
Fixing known issues with running on big endian systems. These changes are focused mainly on ATA devices that have been tested in a big endian system. Other devices, such as NVMe, may require additional changes.
The way this is done was to keep as few changes as possible to the code for now, but a better solution may be found in the future to handle this differently.