quick: Adding a workaround to keep VS2013 compiling
Adding a workaround to keep VS2013 compiling. What happens is VS2013's print format macros do not work in the TEXT macro to expand to wide character strings because of how they are defined. This is not a problem in later versions, so it is hardcoded to work in this situation.