bug: Removing function to set SAT Off-line bits
Removing the function that set the SAT off-line field from the code.
After review, this function was not working correctly and the use of this field is actually for old PATA-HBA interactions. The use of this is better off in a hack/workaround on a per-HBA and/or per-command basis. Since there are no currently known cases for this, it is off until we can find a need for this.