bug: Moving report LUNs to after flags for waking drive and fast scan
Moving the report LUNs code to after flags for waking a drive or the fast scan.
The reason is simple: this changes the drives to active state. Because of this, it violates the DO_NOT_WAKE drive flag, and it placed after FAST_SCAN because drives being scanned while spun down would take 10 or more seconds to spin up, which is not fast.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>