bug: Fixing an issue opening devices and fixing device filters
Fixing a bug where opening a device would fail. There was a missing else for when NVMe is enabled to prevent some potential issues.
The other issue was in some systems partitions are marked with s in the device handle, so the filter needed to be updated to exclude those sine they cannot be opened the same on ata/scsi.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>