feat: Adding device interface lookup for CAM and filtering out devices with slices
Adding in code that properly supports determining the connected interface type, which is most important for USB and firewire.

Also updating the device filters to remove devices that use "s" to denote slices/partitions in addition to the existing "p" filter since this was causing some issues while scanning for devices.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>