feat: Adding function to lock and unlock a device
Adding functions to lock and unlock a device. This is done to try and get exclusive access after calling lock so that other applications cannot interrupt certain commands or operations.
In Windows, this sends some LOCK and UNLOCK IOCTLs, in other systems this removes and adds the O_NONBLOCK flag.
In FreeBSD these are stubs because devices do not get opened with this flag and I don't see any other API calls that do similar things right now.
Currently, the locks/unlocks are meant to be used at a higher level in order to allow wrapping sequences of commands instead of individual commands. This may be explored or changed in the future as needed.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>