bug: Attempting to fix FreeBSD 11.3 compilation with NVMe support
Attempting to fix the FreeBSD 11.3 compilation. Sometime in 11.3 to 11.4 big endian support for NVMe was added and that changed from use a structure for status to a uint16. This means we need to copy the strucutre to a uint16 before using it.
This is currently ifdef'd with a freeBSD version pointing to 11.4 since I do not know the exact version this changed in. It can be more finely tuned in later as needed.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>