feat: Adding a "Basic" passthrough type for unknown devices
Adding a basic device that can handle unknown device types, such as when the AMD ramdisk software is installed.
It relys on Windows IOCTLs alone to function, so it is very limited. It gets treated as a SCSI drive with very little functionality. It might be able to be expanded beyond this, but this is plenty for compatibility at this time.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>