bug: Fixing a pointer addition error in intel fw IOCTL
Fixing pointer addition error that was causing memory corruption and crashes.
The addition was adding 56 * sizeof(IOCTL_RAID_FIRMWARE_BUFFER) rather than only 56 bytes. Casting to a void* first will solve it, but I added a additional member to the structure because this is far more readable and less confusing to figure out.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>