feat: Using versionsort instead of alphasort in sg_helper
Switching to versionsort instead of alphasort in sg_helper. This required _GNU_SOURCE to be defined, which the makefile currently does, but as a backup it also gets defined locally in sg_helper and undefined at the end (if locally defined only).
So this will cause the sg handles to show in numeric order instead of sg1, sg10, sg11, sg2...etc.

This is only done in Linux since this is a GNU extension, but it may be possible to add in FreeBSD code as well since it seems to be listed in newer man pages, but that can be revisited later.

[Fixes #7]

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>