feat: Adding os_Update_File_System_Cache function
Adding a function to try and update file system cache. This has come up as important under windows so that when a disk is erased, then viewed in disk amanger or computer, it can show that it needs formatting. This call is supposed to cause the low-level code to update and pick up file system changes since the os_lock and os_unlock did not work quite as they were documented to.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>