feat: Adding SCSI Miniport Firmware support to windows low-level
Adding support for issuing the older SCSI miniport firmware commands. These can specify timeouts, and return more useful error codes when something goes wrong.
This also should allow firmware updates for NVMe devices in Windows 8.1 that may not have been working before now.

While this code is setup for SATA and SAS, it only seems to work on NVMe devices. The code is left there in case a low-level driver does implement it, it should issue FWDL IOs with this instead of the Win10 API due to the better timeouts and more error information.
On NVMe, the miniport is preffered as well for the same reasons. If a low-level driver doesn't support the miniport IOCTLs, then the standard Windows 10 API code will be used instead.

All miniport code uses the V2 structures for Win10, and V1 structures for Win8.1, but in Win10 both seem to work ok.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>