quick: Renaming REZERO_UNIT to REZERO_UNIT_CMD since it matches a linux sys header
Adding _CMD to REZER_UNIT since it matches a definition in a system header. While I am not currently repeating a compilation issue, we are seeing one in another system and tool using this code which may also be including that header in just the right order to cause an issue.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>