feat: Adding FUA option to generic read/write commands. Removed unused async bool.
Changed the unused async bool to forceUnitAccess (FUA). Async support is much more complicated and was implemented in a separate library since it didn't fit with the "diagnostic" focus that opensea-transport has today.

This  included low-level changes in os_Read and os_Write in Windows that attempts to open a handle with additional flags to make this happen.
If the drive does not support FUA bits/commands, then this is emulated using verify commands. So an emulated read FUA will issue a verify, then a read to force and cached data to the media first.
An emulated write FUA will issue a write, then a verify to force the written data through the cache to media since a verify command requires media access.
Most SATA and SAS drives support this today, so it will likely not be an issue. USB translators do not always support this, and if we are issuing pass-through SATA commands, there is not a way to issue the FPDMA commands that have a separate FUA bit, so emulation is used a lot there as well.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>