bug: Skipping support for the ARCSAS driver in CSMI for now
This change skips supporting the ARCSAS driver with CSMI for now. There is a bug in the code that causes a crash when scanning for devices that has not been fully figured out yet. This is a temporary solution to prevent crashes until the root cause is found.
When the arcsas driver is CSMI enabled (registry key), this code begins scanning it. A customer has 42 Raid Sets and it crashes while scanning the 37th one, but it is not clear why. It could be caused by some other memory error earlier in execution, but for now this will work around the crash until we are able to determine a better solution.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>