feat: Adding another check to make sure USB interface is discovered
Adding another check to make sure a USB device is properly identified as USB. This is more important for systems other than Windows or Linux, but is a good backup in case something goes wrong in those OSs as well.

Also adding another PID for Seagate Firecuda Gaming SSD USB drive that I learned about today.

The overall goal is to try and further help identify NVMe passthrough over USB in case the PID is not yet known.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>