clean: Fixing a warning in code that gets FUA handle in Windows
Fixing a warning in the code that gets the FUA handle since it may not be the correct type to pass to "CreateFile".
It now works as get_Win_Device does and uses an intermediate array to hold the name that is a TCHAR to work whether building in Unicode mode or not.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>