feat: Refactoring DISABLE_NVME_PASSTHROUGH flag
Refactoring the DISABLE_NVME_PASSTHROUGH flag so that it only needs to be used and checked in the low-level code for a given OS.
It really only needs to be in opensea-transport where the low-level interface is available or not, and this is already setup in currently supported compilers for this flag and this resolves that.
It also greatly simplifies the code throughout all the libraries and utilities using it.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>