feat: adding in partition table rescan/update for Linux
Adding in the partition table rescan/update capability to match what happens in Windows. It is recommended that mounted partitions are unmounted first, but that will be another update as that is a little more complicated. This does work properly for FAT and NTFS from what I can tell, but ext4 partitions stick around and cannot be mounted after erasing because it was erased. So there is some other file in the OS tracking this differently from other partitions. If you unmount first, everything works as expected. Still looking up the best way to handle unmounting partitions on the same disk from the OS and that will be added as a second call in another function.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>