feat: Adding in function to find and unmount partitions for a given device
Adding os_Unmount_File_Systems_On_Device(tDevice *device) to  unmount any mounted partitions/filesystems that are currently active on a given device.
Currently, only linux has this fully implemented; all other OSs are stubs for now.
This was added because if you create an ext4 partition on a device, erase it, and call the os_Update_Filesystem_Cache() function, which rescans the partition table, a weird issue happens where the blkid file on the system still thinks the disk should have a partition. You cannot mount the partition because it is destroyed, and it may be tricky to some users to figure out how to correct this issue. By unmounting any partitions before performing an erase, this issue goes away entirely.
I also tested with NTFS and FAT32 partitions and neither of them had the same issue as ext4, but this should prevent this kind of issue regardless of what partitions are being erased so that a drive can be reused without any odd behavior.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>