feat: Adding Windows support for unmounting volumes
Adding support to unmount volumes in Windows for a specified drive.
It works a bit differently than linux as the volumes in use are cached earlier, then that bit field is used again later to unmount devices.
The bitfield used should be for volumes only on the specified device as the bits are added to it when a match is found.
This was done to prevent calling GetLogicalDrives() multiple times as this will remount any volume it can when it is called. If we can find a way to avoid this call in the future, we should make that improvement so that the actions on one disk do not possibly impact another. For now this is good enough.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>