quick: Changing return code to unmount partitions on FreeBSD
Adding code that uses getmntinfo and unmount to find and unmount any mounted partitions on a given device handle. It seems to do what I want from my limited testing in a VM and is similar to what I have done for Linux and Windows so far.
I updated a comment about the os_Update_File_System_Cache function as I haven't found a good way to accomplish the same behavior there. It looks like a MNT_RELOAD may work, but that may only reload the mount file. More testing will be needed before that is implemented.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>