feat: Adding solaris/illumos support for detecting FSs and unmounting them
Adding support for solaris/illumos to unmount file systems that are mounted before things like an erase. Also adding in detection of boot device and when disks have mounted file systems.
NOTE: I do not currently have a system to test this on, but the API was nearly the same as on Linux, so it should work about the same, or be really close.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>