feat: Cleaning up when to set some legacy bits from the oldest ATA specs
Cleaning up when to set some very old device register bits for ATA drives.
The original ATA spec just said "set these to 1" and they have been like that in this code for backwards compatibility for a long time now.
After a talk with Seagate's T13 representative, it was recommended that these bits are not set for SATA drives...only for PATA drives.

This is starting on a separate branch to allow for enough time to test this change before merging into develop.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>