bug: Fixing length and offset for NVMe FWDL in Intel RST helper
Fixing a bug in how the data length and offset are sent to another low-level function in the Intel API for NVMe. Since the NVMe spec reports these as numbers of dwords, a multiplication by 4 is required to fix this issue. This is how it was implemented in the Windows API and was left out of the Intel RST API code on accident.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>