quick: Adding some helper functions to help validate ATA Identify words
Adding functions that help validate that ATA identify words are valid before interpreting the data. The basic version checks if not 0 and not FFFFh. There are two others that validate the state of other bits that the spec requires to be set to zero or 1 for use with certain other words.
Will need to use there in places where this is already being done and will need to update other areas of the code to use these where validity is not currently being checked.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>