feat: Attempting mingw version check before defining some devprop keys
Adding version check to work around failing mingw builds failing. Checking for major version < 10....not sure of a better way to deal with this right now. I have v11 and it has this and v10 in Ubuntu and it gives this error, so hoping this will correct both of those cases.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>