bug: Fixing a crash on some CSMI systems
Fixing a crash that would happen on some CSMI systems, but only sometimes.
The issue was related to the code handling scanning for CSMI raid devices.
As this code runs, it removes any enumerated devices from the raid list. When it removed the last entry in the list, it was not updating a provided pointer back to the OS layer, which then called the delete_RAID_List function on a bad pointer, so a double-free.
After lots of debugging and testing with drmemory this has been fixed and is no longer an issue.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>