clean: Removing last change and replacing with useful comments
The last change added a strange value for a RAID config report, but that was before looking at the full raw memory. Intel's driver is using this previously reserved field for the user defined RAID array name. I added useful comments about this use instead since the value was due to what I named my RAID array.
The last commit was the proper fix for this by using the CSMI minor version number to check when to assess the bDataType field.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>