feat: Fixing CSMI RAID device discovery and adding debug output
Fixing a major issue in the CSMI RAID drive scan. The issue had to do with expecting a SAS address, but that is not guaranteed on SATA controllers. So when these are not valid, it falls back to matching device identify data to make a match instead. This corrects issue I found with Intel's driver and the changes should also be compatible AMD's driver.

Additionally a lot of debug was added that can be turned on or off with a compile time flag if this needs more future debugging.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>