quick: A lot of changes to get things running in AIX
Made a lot of changes, some only clean up warnings, but there were some larger changes that were necessary to get AIX happy to issue commands. Also found out that the IOCINFO did not run on /dev/sasX as expected, so removed relying on that for now. Using the information available I was still able to setup the things we needed to configure to issue commands and it seems to work on the SAS drive I have at the moment, at least for the simple basics.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>