feat: Automatic detection of SC_DIAGNOSTIC flag requirement in AIX
Adding automatic detection for when SC_DIAGNOSTIC is required to access a drive. This was necessary as a normally functioning drive will not need this flag. Also, the boot device will give an EACCESS error with it, but other devices may require this flag.
With these changes, now any device should be able to be enumerated.
I also made some changes to the conditions within get_Device so that the first IOC_INFO is not required since this can be worked around.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>