bug: Working around yet another CSMI issue with RAID config being incomplete
Sometimes requesting the RAID config gives missing MN and SN data, so it is impossible to match against anything in the RAID scan and skip already discovered devices and prevent duplicates. Due to this issue, we can only rely on the PhyInfo which can still find everything, but it may also create an issue where we see an occasional duplicate in the output device list. This is ok for now until we figure out another way around this.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>