feat: Adding code to read the adapter IDs and tested reading mnt info
Wrote code that is able to find the PCI/PCIe adapter IDs for an hdisk and verified the output makes sense, so that has been added.

I attempted to use the mntent.h to figure out mounted partitions. While the code works, I commented it out since I do not know how to figure out which /dev/hd? is on a given /dev/hdisk? which is the missing piece. Modifying this code to work should not be difficult once this has been figured out.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>