feat: Improving CSMI registry key lookup
Improving the CSMI registry key lookup. There are a couple possible places to find the registry key depending on the driver type. This now handles the most likely 2 (miniport and storport). The last possibility is a miniport adapter specific key. I can add this if necessary, but it may be better to find a way to combine it with the current miniport lookup since it will be nearly identical.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>