quick: Adding in reading CSMI controller status to assist with debugging
Adding in reading the controller status. Currently this is just to help dump info in verbose/debug to figure out what is responding. It can be used better to detect a bad controller or a controller in a state where passthrough is not possible in the future. With the known CSMI compatible drivers out there, I'm not sure if we will ever see one in a bad state or not.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>