feat: Adding support to detect maximum time violation case in NVMe fwdl
Detecting when CA 11b returns a maximum time violation and instead it will change to a different activation type that requires a reset. This should improve overall user experience by detecting this. This stays compatible with the forcing of a specific commit action and when disabling the automatic reset after commit.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>