feat: Fixing reading adapter info on AMD RAID and adding reading driver info
Fixed a unique situation with how the AMD RAID driver ends up reporting the adapter information. From the device, the parent of the parent is needed which is different from other configurations I've previously tested.
Additionally I have added reading the driver name and version information so that it can be used or displayed for debugging.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>