feat: Adding driver name string field and some validation bools
Adding some more info to the driver info structure ahead of expanding the linux support to save the full driver version string in case parsing it fails for any reason. Also added some validation booleans to know which parts of the version were actually reported as a quick peek at linux shows a couple different formats that may be returned.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>