doc: Adding sample function for Windows NVMe workaround for reading telemetry
MSDN lists the read buffer 16 CDB as supported for SCSI translation. This basically allows reading the telemetry log, both host and ctrl.
It is unclear which version of Windows first supported this method.

I'm adding a mostly complete function, although commented out, in case we need it in the future. The API to read log pages was updated to allow offsets as well which is the better thing to use.
If we encounter any strange errors trying to pull telemetry on Windows 10, pre 1903, this function may be useful to try using, but we will only want to use it when the other API's ability to specify offsets is not available. 1809 seems to be the version that first made the offsets available, so it is unlikely that an earlier version will be seen, but this may have value in some odd cases.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>