clean: Fixing a signed array subscript warning
Fixing a warning that was reported from GCC in OmniOS. Using a signed value to access and array can potentially be dangerous as it is undefined behavior. This warning has not shown under GCC or Clang in Ubuntu 22.04, but it is still a good warning to make sure is not present in any OS or version of the compiler.
Changed to using unsigned types and changed the loop conditions a little bit. This change should prevent the warning and will break out as expected when we hit the end of the current serial number that is being reversed.

[Seagate/opensea-transport#15]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>