bug: Fixed a write beyond allocated memory in is_LaCie check
Fixed a bug where a null terminator is added to the end of a string, but it was being added past the end of the allocated memory.
This is an old compatibility workaround for snprintf in MSVC prior to VS2015 as that snprintf was not guaranteed to null terminate every string like the standard required.
The hard coded offset was assuming vendor ID length of 8, which is from the standards, but the field being used for this check has already been trimmed of whitespace, so it is more likely to be shorter than the standard 8 bytes.

[Seagate/openSeaChest#99]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>