bug: Setting SCSI version number for NVMe devices
Setting the SCSI version number for NVMe devices. Since these may occasionally use some SCSI translated capabilities or we may want to interact with the SCSI translator, this value is set to what is required in the SCSI translation whitepaper as a safe value for what is most likely accurate. It will only be set if the value is currently zero, so it continues to work with USB attached devices.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>