feat: Adding support for NVM format with SES=crypto or SES=user secure erase
MSFT has some support for performing a NVM format when SES is set to a user secure erase or a crypto secure erase.
This is done through the non-standard, not in SNTL, translation of the SCSI Sanitize command.
I've implemented it and tested it as best I can for now. There is one last thing we can attempt to support, which is using the reinitialize media IOCTL for the crypto erase. This is likely only needed for supporting older versions of Windows 10 and will only work on drives with encryption capabilities, so not a lot of things that will support it. It is also possible that the existing translations work on that older Win10 build, but without one running, there is no way to test it since documentation only goes back to 1903.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>