bug: Fixing the length used in report luns to match minimum from SPC
The report luns command was using 8 bytes to figure out the list length which is fine in new standards, but a problem for older standards.
The SPC spec set the minimum to 16 bytes, so I have made it match at least that to reduce the possibility of errors being reported from this command.
I have also set a requirement that the drive reports a version greater than SCSI 2 to make sure this command will only be issued to devices that are expected to support it.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>