feat: Adding a new OS function to erase the boot sectors (Win only for now)
Adding a new function to call to erase the boot sectors. For Windows, there is an IOCTL that needs to be called on some drives before other writes will be able to pass as a permissions error will be returned otherwise. This call will do this and after it completes, writes to LBA 0 will succeed. It should only be called when actually wanting to erase the boot sectors.
Non-Windows have stubs for this right now since I could not identify any similar IOCTL and I have not noticed this same issue outside of Windows so far.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>