bug: Fixed CSMI scan bug showing some devices twice
There was a bug where scanning in CSMI would occasionally find the same drive twice. After a long time debugging, I removed some of the special-case rules for Intel drivers because they only worked in a situation where all RAID drives were functional and connected.
If a drive was removed, then there are some other things that needed to be checked for. In my debugging I found the RAID config still listed the missing SN, but the MN was blank, so I added a case for that.
I also changed the routine to only check the phy info for the problem cases which stopped finding the same drive twice since the fallback phyinfo scan would always find a device that was already enumerated.
The last thing I changed was to make sure the scan omits ATAPI devices since those will not be in a RAID and do not need to be counted or scanned.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>