feat: Adding initial code to parse ATA rtfrs status and error bitfields
Adding the initial code to parse the status and error bitfield outputs from ATA commands in an effort to further enhance the verbose command output from this software.

This also changes how we parse the returned sense data to not rely only on the ASC and ASCQ being set to "ATA passthrough information available". This needs some more testing before it is finalized, but will help with some odd situations that we have run into on some HBAs when a drive fails particular commands. Sometimes "unaligned write command" is returned, although the fixed format sense data still seems to indicate a valid ATA response like is expected for the "ATA passthrough information available" asc and ascq.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>