feat: Making the SAT workaround for "Unaligned write command" a flagged option
Due to inconsistencies and this being a lot of "guess work" it is difficult if not impossible to know how reliable this workaround is for getting the RTFRs.
It appeared to somewhat work when I tested it, but I do not want to introduce this behavior permanently right now since it is overall unreliable at best and completely wrong at worst.

I walked through the linux kernel code attempting to find exactly where libata would be setting this up in order to validate it, but I cannot find any reason in the libata code or the scsi kernel code that would put the RTFRs in these locations....so I this is being wrapped with a flag so that the code is not lost and can be revisited in the future if we can figure out what would be setting the "unaligned write command" sense with possible rtfrs of value.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>