bug: Fixing a bug when trying to read the current NVM power state in Windows
In Windows, the IOCTL to get the current feature value expects some amount of data transfer, even though an NVMe get features command may transfer no data at all.
The workaround to this was to set 4 bytes when the command request is set to zero.
I chose 4 bytes since most of NVMe works in DWords.

This fix is not ideal as Windows is apparently filtering the set features request to force using their other IOCTLs used in previous versions of Windows 10 so it has a retry in here as well. I am keeping this since the set features is more feature rich to allow more fields to be set if it is ever accepted for the power state feature in the future.

Signed-off-by: Tyler Erickson <tyler.j.erickson@seagate.com>