feat: Initial code changes to support realtek 9210 USB adapter
Checking in initial changes that make supporting the realtek 9210 possible.
This device is pretty different in behavior from previous USB to NVMe chips since it can also support SATA M.2 devices.
Because of this, it required a bit more testing and tweaking to figure out the initial capabilities.
The most peculiar quirk is that when an NVMe drive is attached, it will respond to an ATA identify command, however this seems to be the only command to respond and the only information filled in is the MN, SN, and FWRev.

There is still more work to do for this since it may be possible to further extend capabilities, but this first set of changes helps a lot with compatibility and performance (at least getting drive info).

[Seagate/openSeaChest#109]

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>